支付与跟团游衔接2

+++ src/main/java/com/hm/travel/pojo/TourOrder.java
@@ -34,9 +34,9 @@ public class TourOrder {
 
     private Traveller traveller;
 
-    private int travellerPrice;
+    private double travellerPrice;
     /**
      * 产品单价
      */
-    private int travellerOneprice;
+    private double travellerOneprice;
 }
